requirements.txt
tableauserverclient>=0.31
python-dotenv>=1.0.1
